+++ apps/client/src/features/review/components/index.ts
@@ -1,0 +1,1 @@
+export { ReviewDeleteConfirmDialog } from './reviewDeleteConfirmDialog/ReviewDeleteConfirmDialog'
