+++ apps/client/src/pages/myReviews/components/WrittenReviewCard.tsx
@@ -1,7 +1,8 @@
-import { MenuIcon, OrderCancelIcon } from '@hashi/hds-icons'
+import { MenuIcon } from '@hashi/hds-icons'
 import { Dialog, StarRating } from '@hashi/hds-ui'
 import { useEffect, useId, useRef, useState } from 'react'
 
+import { ReviewDeleteConfirmDialog } from '@/features/review/components'
 import { ReviewImagePlaceholder } from '@/pages/myReviews/components/ReviewImagePlaceholder'
 import type { WrittenReview } from '@/pages/myReviews/types/myReview'
 
@@ -13,12 +14,6 @@ interface WrittenReviewCardProps {
   onEdit: () => void
   onToggleMenu: () => void
 }
-
-const secondaryActionClassName =
-  'typo-body-6 flex h-9 items-center justify-center rounded-[5px] bg-secondary-200 text-black'
-
-const primaryActionClassName =
-  'typo-body-6 flex h-9 items-center justify-center rounded-[5px] bg-cool-gray-800 text-white'
 
 export const WrittenReviewCard = ({
   isMenuOpen,
@@ -155,33 +150,5 @@ const ReviewMoreMenu = ({ id, onDelete, onEdit }: ReviewMoreMenuProps) => {
         삭제하기
       </button>
     </div>
-  )
-}
-
-export const ReviewDeleteConfirmDialog = ({
-  onDelete,
-}: Pick<WrittenReviewCardProps, 'onDelete'>) => {
-  return (
-    <Dialog.Content className="py-[23px]">
-      <Dialog.Header>
-        <Dialog.Icon>
-          <OrderCancelIcon className="size-6" />
-        </Dialog.Icon>
-        <div className="flex flex-col items-center gap-2">
-          <Dialog.Title>정말 삭제하시겠습니까?</Dialog.Title>
-          <Dialog.Description>
-            삭제한 리뷰는 다시 되돌릴 수 없어요.
-          </Dialog.Description>
-        </div>
-      </Dialog.Header>
-      <Dialog.Footer>
-        <Dialog.Close className={secondaryActionClassName}>
-          취소하기
-        </Dialog.Close>
-        <Dialog.Close className={primaryActionClassName} onPress={onDelete}>
-          삭제하기
-        </Dialog.Close>
-      </Dialog.Footer>
-    </Dialog.Content>
   )
 }
